azure_functions/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.14.3
azure-identity==1.25.1
